refactor : move character selection to core:ui
