+++ app/src/main/kotlin/com/chipichipi/dobedobe/onboarding/OnboardingSelectCharacterScreen.kt
@@ -1,23 +1,10 @@
 package com.chipichipi.dobedobe.onboarding
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.chipichipi.dobedobe.R
-import com.chipichipi.dobedobe.core.designsystem.component.DobeDobeTextButton
-import com.chipichipi.dobedobe.core.designsystem.theme.DobeDobeTheme
-import com.chipichipi.dobedobe.core.model.CharacterType
+import com.chipichipi.dobedobe.core.ui.SelectCharacterScreen
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
@@ -27,60 +14,10 @@ internal fun OnboardingSelectCharacterRoute(
 ) {
     val selectedCharacter by viewModel.selectedCharacter.collectAsStateWithLifecycle()
 
-    OnboardingSelectCharacterScreen(
+    SelectCharacterScreen(
         modifier = modifier,
         selectedCharacter = selectedCharacter,
         onCharacterToggled = viewModel::toggleCharacter,
-        onOnboardingCompleted = viewModel::completeOnboarding,
+        onCompleted = viewModel::completeOnboarding,
     )
-}
-
-@Composable
-private fun OnboardingSelectCharacterScreen(
-    selectedCharacter: CharacterType,
-    onCharacterToggled: () -> Unit,
-    onOnboardingCompleted: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(horizontal = 24.dp)
-            .padding(
-                top = 48.dp,
-                bottom = 32.dp,
-            ),
-    ) {
-        Text(
-            text = stringResource(R.string.onboarding_select_character_title),
-            style = DobeDobeTheme.typography.heading1,
-            color = DobeDobeTheme.colors.black,
-        )
-        Spacer(modifier = Modifier.height(48.dp))
-
-        OnboardingBirdCard(
-            selected = selectedCharacter == CharacterType.Bird,
-            onClick = onCharacterToggled,
-        )
-        Spacer(modifier = Modifier.height(32.dp))
-
-        OnboardingRabbitCard(
-            selected = selectedCharacter == CharacterType.Rabbit,
-            onClick = onCharacterToggled,
-        )
-        Spacer(modifier = Modifier.weight(1f))
-
-        DobeDobeTextButton(
-            modifier = Modifier
-                .fillMaxWidth()
-                .heightIn(56.dp),
-            onClick = onOnboardingCompleted,
-        ) {
-            Text(
-                text = stringResource(R.string.onboarding_all_complete),
-                style = DobeDobeTheme.typography.heading2,
-                color = DobeDobeTheme.colors.white,
-            )
-        }
-    }
 }
