+++ app/src/main/res/values/strings.xml
@@ -5,9 +5,4 @@
     <string name="onboarding_goal_hint">ex) 1일 1책 읽기</string>
     <string name="onboarding_goal_support_message">목표가 간결할수록 집중력이 높아져요.</string>
     <string name="onboarding_goal_completed">작성 완료</string>
-
-    <string name="onboarding_select_character_title">함께 할 러닝메이트를\n선택해 주세요</string>
-    <string name="onboarding_all_complete">시작하기</string>
-    <string name="onboarding_bird_character_description">마구 쪼는 잔소리무새</string>
-    <string name="onboarding_rabbit_character_description">공감 만점 럭키 토끼</string>
 </resources>
