+++ core/ui/src/main/res/drawable/bird_character.png
