+++ core/ui/src/main/res/drawable/bubble_bird.png
