+++ core/ui/src/main/res/drawable/bubble_rabbit.png
