+++ core/ui/src/main/res/drawable/rabbit_character.png
